pyproject.toml
[project]
name = "voice-to-text"
version = "0.1.0"
description = "Real-time speech-to-text transcription service"
requires-python = ">=3.10"
dependencies = [
    # FastAPI and web server
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.0",
    "websockets>=14.0",

    # NeMo and ML dependencies for ASR
    "nemo-toolkit[asr]>=2.6.0",
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "numpy>=2.0.0",
    "librosa>=0.11.0",
    "soundfile>=0.13.0",
    "scipy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
]
